Implemented Faraway Island, added GCN location data, updated Fldeff_Cut
This is a big one. Faraway island has been fully implemented and is currently accessible from the Seagallop Ferry menu as a proof of concept. The island works fully as it does in Emerald, and several new tiles and other data had to be added to accommodate this.

Additionally, location data has been added for both the GCN games and Johto (which will serve as the basis of my own project), and trading with ANY compatible game from the GBA or GCN will now show its met location on the summary screen.

Lastly, Fldeff_Cut has been fully replaced with Emerald's version, essentially still cosmetically identical to FireRed's--but now with working support for both long grass and hyper cutter!