Added RTC and time-based evolutions to the game, updated save loading function
Implemented the RTC and the RTC reset screen, as well as the ability for Eevee to evolve in to either Espeon or Umbreon based on the time of day. Unlike in Ruby, Sapphire, and Emerald, the time periods necessary to do so are reflective of Gen 2's day and night periods. Additionally, save file loading has been moved earlier in the init process so that corruption caused by the RTC reset function will no longer occur.

Lastly, I've also added the graphics for the Cushion decorations to the repository, but they are not yet integrated in to the game in any capacity.