Wild encounters can now vary by time of day
Added morning/day/night data for wild land encounters, allowing for unique data to be assigned for all maps based on the time of day. Altering Cave is given a special exception and only uses its normal encounter data as per usual. Additionally, a "hotfix" was added which swaps out any instance of a wild Corsola while fishing with that of Staryu at night--as Corsola/Staryu were the only fishing-based Pokemon who varied based on time of day in GSC, it felt pointless to make new encounter slots just for this singular occurrence.