Added ability to plant and grow berries
Added the ability to plant and grow berry trees. This works as it does in Emerald. The berry pouch has been updated to reflect this new functionality. The Wailmer Pail now functions as it did in RSE, and the Squirtbottle from Gens 2 and 4 has been added with identical functionality. The ability to "water" a Sudowoodo has been partially implemented, but the remainder of its functionality depends on the existence of a specific event.